+++ src/main.rs
@@ -1,6 +1,7 @@
 mod ast;
 mod bytecode;
 mod compiler;
+mod error;
 mod heap;
 mod value;
 mod vm;
@@ -33,8 +34,9 @@ fn main() {
         eprintln!("{:#?}", prog);
     }
 
+    let filename: std::rc::Rc<str> = std::rc::Rc::from(args[1].clone());
     let mut protos: Vec<Proto> = vec![];
-    let entry = compile_proto("<main>".into(), vec![], &[prog], &mut protos);
+    let entry = compile_proto("<main>".into(), vec![], &[prog], filename, &mut protos);
     if env::var("DEBUG_BC").is_ok() {
         for (i, p) in protos.iter().enumerate() {
             eprintln!("proto {} {}", i, p.name);